add a filter to FilteredExamData
