+++ exam_graph/utils/exam_data_set.py
@@ -5,50 +5,20 @@ import pandas as pd
 from exam_graph import helper
 from exam_graph import filters
 from pathlib import Path
-from datetime import datetime, date
+from datetime import datetime, date, time
 from formatter import *
 import validator
 
 
-class HL7Fields:
 
-    def __init__(self):
-        self.order_complete_dt = 'Exam Complete Date/Tm' # this is our ORM.ORC.5 (order status is complete)
-        self.acc = 'Order Procedure Accession' # ORM.ORC.2
-        self.order_dt = 'Exam Order Date/Time' # ex:'2024-07-10T01:15:00'; this is our HL7:ORM.ORC.15 (NW order time)
-        self.final_dt = 'Final Date/Tm' # this is our ORM.OBR.22.1 (time of report)
-        self.order_name = 'Exam Order Name' # ORM.OBR.4
-        self.modality = 'Modality'
-        
-    def get_fields(self) -> dict:
-        return {
-            'ORC5': self.order_complete_dt,
-            'ORC2': self.acc,
-            'ORC15': self.order_dt,
-            'OBR22_1': self.final_dt,
-            'OBR4': self.order_name,
-            'OBR24': self.modality
-        }
-    
-    def get_columns(self) -> list:
-        return  list(self.get_fields().values())
-    
-    # returns a df with attributes as columns and no data
-    def make_df(self) -> pd.DataFrame:
-        fields = self.get_columns(self)
-        df = pd.DataFrame(columns=fields)
-        return df
-    
-    @classmethod
-    def show_default_fields(cls) -> list:
-        return  list(cls().get_fields().values())
     
 # csv gets ingested into this class
 # we can do validation here and df conversion
 class ExamDataFrame:
-    def __init__(self, df:pd.DataFrame=None, hl7_fields=HL7Fields()):
+    def __init__(self, df:pd.DataFrame=None, hl7_fields=HL7Fields(), shifts=Shifts().dict()):
         self.hl7_fields = hl7_fields
         self.df = None
+        self.shifts = shifts
 
     def read_file(self, fp:Path):
         accepted_file_types = {'.csv': pd.read_csv}
@@ -58,7 +28,7 @@ class ExamDataFrame:
             self.df = accepted_file_types[file_type](self.fp)
             self.df.name = fp.stem
             print(f'Successfully read {self.fp.stem} to df',end='\n\n')
-            self.df = Formatter(df=self.df,hl7=self.hl7_fields).format_df()
+            self.df = Formatter(df=self.df, shifts=self.shifts, hl7=self.hl7_fields).format_df()
         else:
             print('Cannot read file',end='\n\n')
 
@@ -138,10 +108,36 @@ class UserFilters:
 
 # validated df goes here and a filter
 # this is where we apply all the filters
-class FilteredExamData(ExamDataFrame):
-    def __init__(self, df:pd.DataFrame, filter:str):
-        super().__init__(df)
-        self.filter = filter
+class FilteredExamData():
+    def __init__(self, exam_data:ExamDataFrame):
+        self.exam_data = exam_data
+        self.df = pd.DataFrame(exam_data.get_df())
+        self.hl7_fields = exam_data.hl7_fields
+
+
+
+    def period(self, period_selection:str) -> pd.DataFrame:
+        # if df == None:
+        #     df = self.df
+
+
+        # use pandas time period aliases for period_selection 
+        alias = {
+            'hour': 'H',
+            'day': 'D',
+            'week':'W',
+            'month': 'M',
+            'year': 'Y'
+        }
+
+        # map user's period selection to pandas period alias
+        period_selection = period_selection.lower()
+        period_selection = alias[period_selection]
+
+        # create new column to we can group by the user's selected period
+        self.df['User_selected_period'] = self.df[self.hl7_fields.order_complete_dt].dt.to_period(period_selection)
+
+        return self.df    
 
     
 
@@ -154,6 +150,16 @@ def main():
     master_df = ExamDataFrame()
     master_df.read_file(my_csv)
     df = master_df.get_df()
+
+    print(df.head())
+
+    filt = FilteredExamData(master_df)
+
+
+    period_filtered= filt.period('Week')
+    print(period_filtered.head())
+    # filt = filt.period('Week').head()
+    # print(filt.head())
     # df['Modality'] = 1
 
     # master_df.validate_columns()
