+++ exam_graph/utils/formatter.py
@@ -1,7 +1,41 @@
 import pandas as pd
 from datetime import datetime, time
 import traceback
-from exam_graph.utils.exam_data_set import HL7Fields
+
+
+class HL7Fields:
+
+    def __init__(self):
+        self.order_complete_dt = 'Exam Complete Date/Tm' # this is our ORM.ORC.5 (order status is complete)
+        self.acc = 'Order Procedure Accession' # ORM.ORC.2
+        self.order_dt = 'Exam Order Date/Time' # ex:'2024-07-10T01:15:00'; this is our HL7:ORM.ORC.15 (NW order time)
+        self.final_dt = 'Final Date/Tm' # this is our ORM.OBR.22.1 (time of report)
+        self.order_name = 'Exam Order Name' # ORM.OBR.4
+        self.modality = 'Modality'
+        
+    def get_fields(self) -> dict:
+        return {
+            'ORC5': self.order_complete_dt,
+            'ORC2': self.acc,
+            'ORC15': self.order_dt,
+            'OBR22_1': self.final_dt,
+            'OBR4': self.order_name,
+            'OBR24': self.modality
+        }
+    
+    def get_columns(self) -> list:
+        return  list(self.get_fields().values())
+    
+    # returns a df with attributes as columns and no data
+    def make_df(self) -> pd.DataFrame:
+        fields = self.get_columns(self)
+        df = pd.DataFrame(columns=fields)
+        return df
+    
+    @classmethod
+    def show_default_fields(cls) -> list:
+        return  list(cls().get_fields().values())
+    
 
 class Shifts:
     def __init__(self):
